chore: Moved interceptor and guard to auth folder becasue they are both auth-related
