+++ src/frontend/src/app/app.config.ts
@@ -3,7 +3,7 @@ import { provideRouter } from '@angular/router';
 
 import { routes } from './app.routes';
 import { provideHttpClient, withInterceptors } from '@angular/common/http';
-import { authInterceptor } from './core/interceptors/auth-interceptor';
+import { authInterceptor } from './core/auth/interceptors/auth-interceptor';
 
 export const appConfig: ApplicationConfig = {
   providers: [
