+++ src/frontend/src/app/core/auth/guards/auth-guard.ts
@@ -2,9 +2,9 @@ import { HttpClient } from '@angular/common/http';
 import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
 import { catchError, map, of } from 'rxjs';
-import { apiUrl } from '../../../environments/environment';
+import { apiUrl } from '../../../../environments/environment';
 
-export const authGuard: CanActivateFn = (route, state) => {
+export const authGuard: CanActivateFn = () => {
   const baseUrl = `${apiUrl}/api/profile`;
   const http = inject(HttpClient);
   const router = inject(Router);
